# watchdog/openapi.yaml
openapi: 3.1.0
info:
  title: Watchdog with Your Infrastructure
  description: Unified API for managing Tempo, Loki, AlertManager, Grafana, and Resolver
  version: 0.0.1
paths:
...
  /api/auth/api-keys:
    get:
      tags:
      - authentication
      summary: List Api Keys
      operationId: list_api_keys_api_auth_api_keys_get
      security:
      - HTTPBearer: []
      parameters:
      - name: show_hidden
        in: query
        required: false
        schema:
          type: string
          pattern: ^(true|false)$
          default: 'false'
          title: Show Hidden
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
    post:
      tags:
      - authentication
      summary: Create Api Key
      operationId: create_api_key_api_auth_api_keys_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
  /api/auth/api-keys/{key_id}:
    patch:
      tags:
      - authentication
      summary: Update Api Key
      operationId: update_api_key_api_auth_api_keys__key_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
    delete:
      tags:
      - authentication
      summary: Delete Api Key
      operationId: delete_api_key_api_auth_api_keys__key_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/api-keys/{key_id}/otlp-token/regenerate:
    post:
      tags:
      - authentication
      summary: Regenerate Api Key Otlp Token
      operationId: regenerate_api_key_otlp_token_api_auth_api_keys__key_id__otlp_token_regenerate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/api-keys/{key_id}/hide:
    post:
      tags:
      - authentication
      summary: Hide Api Key
      operationId: hide_api_key_api_auth_api_keys__key_id__hide_post
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HideTogglePayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/api-keys/{key_id}/shares:
    get:
      tags:
      - authentication
      summary: Get Api Key Shares
      operationId: get_api_key_shares_api_auth_api_keys__key_id__shares_get
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyShareUser'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
    put:
      tags:
      - authentication
      summary: Put Api Key Shares
      operationId: put_api_key_shares_api_auth_api_keys__key_id__shares_put
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyShareUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyShareUser'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
  /api/auth/api-keys/{key_id}/shares/{shared_user_id}:
    delete:
      tags:
      - authentication
      summary: Remove Api Key Share
      operationId: remove_api_key_share_api_auth_api_keys__key_id__shares__shared_user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Key Id
      - name: shared_user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Shared User Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
...
  /api/auth/mfa/enroll:
    post:
      tags:
      - authentication
      summary: Mfa Enroll
      operationId: mfa_enroll_api_auth_mfa_enroll_post
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotpEnrollResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
      security:
      - HTTPBearer: []
  /api/auth/mfa/verify:
    post:
      tags:
      - authentication
      summary: Mfa Verify
      operationId: mfa_verify_api_auth_mfa_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
      security:
      - HTTPBearer: []
  /api/auth/mfa/disable:
    post:
      tags:
      - authentication
      summary: Mfa Disable
      operationId: mfa_disable_api_auth_mfa_disable_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaDisableRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
      security:
      - HTTPBearer: []
  /api/auth/users/{user_id}/mfa/reset:
    post:
      tags:
      - authentication
      summary: Admin Reset User Mfa
      operationId: admin_reset_user_mfa_api_auth_users__user_id__mfa_reset_post
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: User Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
...
components:
  schemas:
...
    ApiKey:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          pattern: ^[^\x00-\x1F]+$
          title: Name
        id:
          type: string
          title: Id
        key:
          type: string
          title: Key
        otlp_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Otlp Token
          description: Secure OTLP ingest token for gateway authentication
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        owner_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Username
        is_shared:
          type: boolean
          title: Is Shared
          default: false
        can_use:
          type: boolean
          title: Can Use
          default: true
        shared_with:
          items:
            $ref: '#/components/schemas/ApiKeyShareUser'
          type: array
          title: Shared With
        is_default:
          type: boolean
          title: Is Default
          default: false
        is_enabled:
          type: boolean
          title: Is Enabled
          default: true
        is_hidden:
          type: boolean
          title: Is Hidden
          default: false
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
      type: object
      required:
      - name
      - id
      - key
      - created_at
      title: ApiKey
    ApiKeyCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          pattern: ^[^\x00-\x1F]+$
          title: Name
        key:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 3
            pattern: ^[^\x00-\x1F]+$
          - type: 'null'
          title: Key
          description: Optional custom API key value (org_id / X-Scope-OrgID)
      type: object
      required:
      - name
      title: ApiKeyCreate
...
    ApiKeyShareUpdateRequest:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        group_ids:
          items:
            type: string
          type: array
          title: Group Ids
      type: object
      title: ApiKeyShareUpdateRequest
    ApiKeyShareUser:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        can_use:
          type: boolean
          title: Can Use
          default: true
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - user_id
      - created_at
      title: ApiKeyShareUser
    ApiKeyUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        is_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Enabled
        is_default:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Default
      type: object
      title: ApiKeyUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HideTogglePayload:
      properties:
        hidden:
          type: boolean
          title: Hidden
          default: true
      type: object
      title: HideTogglePayload
...
    MfaDisableRequest:
      properties:
        current_password:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Password
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      type: object
      title: MfaDisableRequest
    MfaVerifyRequest:
      properties:
        code:
          type: string
          title: Code
      type: object
      required:
      - code
      title: MfaVerifyRequest
...
    RecoveryCodesResponse:
      properties:
        recovery_codes:
          items:
            type: string
          type: array
          title: Recovery Codes
      type: object
      required:
      - recovery_codes
      title: RecoveryCodesResponse
...
    TotpEnrollResponse:
      properties:
        otpauth_url:
          type: string
          title: Otpauth Url
        secret:
          type: string
          title: Secret
      type: object
      required:
      - otpauth_url
      - secret
      title: TotpEnrollResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    JSONValue:
      anyOf: *id002
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    WatchdogCookieAuth:
      type: apiKey
      in: cookie
      name: watchdog_token
      description: Authenticate using the watchdog_token cookie (same token as Bearer).
...